Pylint and test
