+++ tests/drawer/test_draw.py
@@ -33,108 +33,6 @@ def circuit(x, y, z):
     return qml.expval(qml.PauliZ(0))
 
 
-@pytest.mark.external
-class TestCatalyst:
-    """Drawer integration test with Catalyst jitted QNodes."""
-
-    def test_simple_circuit(self):
-        """Test a simple circuit that does not use Catalyst features."""
-        import catalyst
-
-        @catalyst.qjit
-        @qml.qnode(qml.device("lightning.qubit", wires=(0, "a", 1.234)))
-        def circuit(x, y, z):
-            """A quantum circuit on three wires."""
-            qml.RX(x, wires=0)
-            qml.RY(y, wires="a")
-            qml.RZ(z, wires=1.234)
-            return qml.expval(qml.PauliZ(0))
-
-        expected = "    0: ──RX─┤  <Z>\n    a: ──RY─┤     \n1.234: ──RZ─┤     "
-        assert draw(circuit, decimals=None)(1.234, 2.345, 3.456) == expected
-
-    @pytest.mark.parametrize("c", [0, 1])
-    def test_cond_circuit(self, c):
-        """Test a circuit with a Catalyst conditional."""
-        import catalyst
-
-        @catalyst.qjit
-        @qml.qnode(qml.device("lightning.qubit", wires=(0, "a", 1.234)))
-        def circuit(x, y, z, c):
-            """A quantum circuit on three wires."""
-
-            @catalyst.cond(c)
-            def conditional_flip():
-                qml.PauliX(wires=0)
-
-            qml.RX(x, wires=0)
-            conditional_flip()
-            qml.RY(y, wires="a")
-            qml.RZ(z, wires=1.234)
-            return qml.expval(qml.PauliZ(0))
-
-        expected = [
-            "    0: ──RX─┤  <Z>\n    a: ──RY─┤     \n1.234: ──RZ─┤     ",
-            "    0: ──RX──X─┤  <Z>\n    a: ──RY────┤     \n1.234: ──RZ────┤     ",
-        ]
-        assert draw(circuit, decimals=None)(1.234, 2.345, 3.456, c) == expected[c]
-
-    @pytest.mark.parametrize("c", [1, 2])
-    def test_for_loop_circuit(self, c):
-        """Test a circuit with a Catalyst for_loop"""
-        import catalyst
-
-        @catalyst.qjit
-        @qml.qnode(qml.device("lightning.qubit", wires=3))
-        def circuit(x, y, z, c):
-            """A quantum circuit on three wires."""
-
-            @catalyst.for_loop(0, c, 1)
-            def loop(i):
-                qml.Hadamard(wires=i)
-
-            qml.RX(x, wires=0)
-            loop()
-            qml.RY(y, wires=1)
-            qml.RZ(z, wires=2)
-            return qml.expval(qml.PauliZ(0))
-
-        expected = [
-            "0: ──RX──H─┤  <Z>\n1: ──RY────┤     \n2: ──RZ────┤     ",
-            "0: ──RX──H──┤  <Z>\n1: ──H───RY─┤     \n2: ──RZ─────┤     ",
-        ]
-        assert draw(circuit, decimals=None)(1.234, 2.345, 3.456, c) == expected[c - 1]
-
-    @pytest.mark.parametrize("c", [0, 1])
-    def test_while_loop_circuit(self, c):
-        """Test a circuit with a Catalyst while_loop"""
-        import catalyst
-
-        @catalyst.qjit
-        @qml.qnode(qml.device("lightning.qubit", wires=3))
-        def circuit(x, y, z, c):
-            """A quantum circuit on three wires."""
-
-            @catalyst.while_loop(lambda x: x < 2.0)
-            def loop_rx(x):
-                # perform some work and update (some of) the arguments
-                qml.RX(x, wires=0)
-                return x + 1
-
-            # apply the while loop
-            qml.RX(x, wires=0)
-            qml.RY(y, wires=1)
-            loop_rx(c)
-            qml.RZ(z, wires=2)
-            return qml.expval(qml.PauliZ(0))
-
-        expected = [
-            "0: ──RX──RX──RX─┤  <Z>\n1: ──RY─────────┤     \n2: ──RZ─────────┤     ",
-            "0: ──RX──RX─┤  <Z>\n1: ──RY─────┤     \n2: ──RZ─────┤     ",
-        ]
-        assert draw(circuit, decimals=None)(1.234, 2.345, 3.456, c) == expected[c]
-
-
 class TestLabelling:
     """Test the wire labels."""
